Refactoring things for multi asset withdrawing from Kraken

+++ src/Model/CompletedWithdraw.php
@@ -8,9 +8,9 @@ class CompletedWithdraw
 {
     protected string $id;
     protected string $recipientAddress;
-    protected int $netAmount;
+    protected float $netAmount;
 
-    public function __construct(string $recipientAddress, int $netAmount, string $id)
+    public function __construct(string $recipientAddress, float $netAmount, string $id)
     {
         $this->id = $id;
         $this->recipientAddress = $recipientAddress;
@@ -22,7 +22,7 @@ class CompletedWithdraw
         return $this->recipientAddress;
     }
 
-    public function getNetAmount(): int
+    public function getNetAmount(): float
     {
         return $this->netAmount;
     }
